fix(python-client): address CI failure and review comments
- Fix CI mypy failure: ScheduleTransitionStruct.transitionTime expects
  the SDK uint type, not a plain int.
- Use cached_fields(type(obj_in)) instead of fields(obj_in): the cache
  key must be the (hashable) class, not the instance, which can be
  unhashable for dataclasses holding list fields (e.g. Schedule with a
  transitions list) -- fields()/cached_fields() accept either the class
  or an instance and return the same result either way, so this is a
  pure caching fix, not a behavior change.
- Clarify the docstring: the name-keyed fallback only applies to
  dataclasses missing CHIP's `descriptor` metadata, which real cluster
  structs (the only values write_attribute() is ever called with)
  always provide.

+++ python_client/tests/test_util.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 from chip.clusters import Objects as clusters
 from chip.clusters.Types import NullValue
+from chip.tlv import uint
 from matter_server.common.helpers.util import dataclass_to_dict, dataclass_to_tag_dict
 
 
@@ -48,7 +49,7 @@ def test_dataclass_to_tag_dict_is_recursive_through_lists() -> None:
         transitions=[
             clusters.Thermostat.Structs.ScheduleTransitionStruct(
                 dayOfWeek=clusters.Thermostat.Bitmaps.ScheduleDayOfWeekBitmap.kMonday,
-                transitionTime=420,
+                transitionTime=uint(420),
                 heatingSetpoint=2100,
             )
         ],
